# backend/schemas/openapi.json
{
  "components": {
    "schemas": {
      "AcceptanceResultResponse": {
        "description": "One criterion of `design.md` \u00a78's ``awaiting_review`` checklist.",
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "created_ms": {
            "title": "Created Ms",
            "type": "integer"
          },
          "criterion": {
            "title": "Criterion",
            "type": "string"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/CriterionOutcome"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "updated_ms": {
            "title": "Updated Ms",
            "type": "integer"
          }
        },
        "required": [
          "node_id",
          "attempt",
          "position",
          "criterion",
          "outcome",
          "created_ms",
          "updated_ms"
        ],
        "title": "AcceptanceResultResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MergeResponse": {
        "properties": {
          "commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit"
          },
          "conflicts": {
            "items": {
              "format": "path",
              "type": "string"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "commit",
          "conflicts"
        ],
        "title": "MergeResponse",
        "type": "object"
      },
      "NodeResponse": {
        "properties": {
          "acceptance_criteria": {
            "items": {
              "type": "string"
            },
            "title": "Acceptance Criteria",
            "type": "array"
          },
          "base_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Ref"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "created_ms": {
            "title": "Created Ms",
            "type": "integer"
          },
          "estimated_effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Effort"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/NodeStatus"
          },
          "touches": {
            "items": {
              "type": "string"
            },
            "title": "Touches",
            "type": "array"
          },
          "updated_ms": {
            "title": "Updated Ms",
            "type": "integer"
          },
          "worktree_path": {
            "anyOf": [
              {
                "format": "path",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worktree Path"
          }
        },
        "required": [
          "id",
          "session_id",
          "name",
          "prompt",
          "acceptance_criteria",
          "harness",
          "model",
          "touches",
          "estimated_effort",
          "worktree_path",
          "branch",
          "base_ref",
          "status",
          "created_ms",
          "updated_ms"
        ],
        "title": "NodeResponse",
        "type": "object"
      },
      "NodeReviewResponse": {
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "decision": {
            "$ref": "#/components/schemas/ReviewDecision"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "reviewed_ms": {
            "title": "Reviewed Ms",
            "type": "integer"
          }
        },
        "required": [
          "node_id",
          "attempt",
          "decision",
          "feedback",
          "reviewed_ms"
        ],
        "title": "NodeReviewResponse",
        "type": "object"
      },
...
      "RejectRequest": {
        "additionalProperties": false,
        "properties": {
          "feedback": {
            "minLength": 1,
            "title": "Feedback",
            "type": "string"
          },
          "outcomes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CriterionOutcome"
            },
            "title": "Outcomes",
            "type": "object"
          }
        },
        "required": [
          "feedback"
        ],
        "title": "RejectRequest",
        "type": "object"
      },
      "RetryRequest": {
        "additionalProperties": false,
        "description": "A new attempt at a ``failed`` or ``blocked`` node.\n\n``feedback`` is optional here and mandatory on reject, and that asymmetry is\nthe orchestrator's: retry is \"try again\", reject is a human overruling a\nfinished attempt.",
        "properties": {
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          }
        },
        "title": "RetryRequest",
        "type": "object"
      },
...
      "ReviewOutcomesRequest": {
        "additionalProperties": false,
        "description": "The reviewer's answers to the acceptance checklist, by position.\n\nPartial on purpose: a position left out keeps whatever it had, which is\n``unevaluated`` until somebody says otherwise. The orchestrator decides\nwhether a ``fail`` disqualifies an approval \u2014 it does not, deliberately\n(see :meth:`~app.orchestrator.service.NodeRunService.approve_node`) \u2014 and\nthis transport does not second-guess it.",
        "properties": {
          "outcomes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CriterionOutcome"
            },
            "title": "Outcomes",
            "type": "object"
          }
        },
        "title": "ReviewOutcomesRequest",
        "type": "object"
      },
...
      "RunResponse": {
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "created_ms": {
            "title": "Created Ms",
            "type": "integer"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "finished_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished Ms"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "harness_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harness Session Id"
          },
          "harness_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harness Version"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "permission_denial_count": {
            "title": "Permission Denial Count",
            "type": "integer"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "started_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started Ms"
          },
          "status": {
            "$ref": "#/components/schemas/RunState"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "id",
          "node_id",
          "session_id",
          "attempt",
          "status",
          "harness",
          "model",
          "pid",
          "harness_session_id",
          "harness_version",
          "started_ms",
          "finished_ms",
          "exit_code",
          "summary",
          "event_count",
          "permission_denial_count",
          "created_ms"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AgentHub",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sessions/{session_id}/node": {
      "get": {
        "operationId": "get_node_api_sessions__session_id__node_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Node",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{session_id}/nodes": {
      "get": {
        "operationId": "list_nodes_api_sessions__session_id__nodes_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NodeResponse"
                  },
                  "title": "Response List Nodes Api Sessions  Session Id  Nodes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Nodes",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/sessions/{session_id}/nodes/{node_id}": {
      "get": {
        "operationId": "get_node_api_sessions__session_id__nodes__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Node",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/sessions/{session_id}/nodes/{node_id}/acceptance": {
      "get": {
        "description": "The per-criterion checklist, oldest attempt first.\n\nWithout ``attempt`` this is the whole history, which is what a drawer\nshowing \"attempt 2 fixed the criterion attempt 1 failed\" needs. With it,\none attempt \u2014 the panel `design.md` \u00a78 specifies for ``awaiting_review``.",
        "operationId": "list_acceptance_results_api_sessions__session_id__nodes__node_id__acceptance_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "attempt",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Attempt"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AcceptanceResultResponse"
                  },
                  "title": "Response List Acceptance Results Api Sessions  Session Id  Nodes  Node Id  Acceptance Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Acceptance Results",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/sessions/{session_id}/nodes/{node_id}/approve": {
      "post": {
        "description": "The human gate's yes: record the verdict, then merge (invariant 6).",
        "operationId": "approve_node_api_sessions__session_id__nodes__node_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReviewOutcomesRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Node",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/sessions/{session_id}/nodes/{node_id}/kill": {
      "post": {
        "operationId": "kill_node_api_sessions__session_id__nodes__node_id__kill_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kill Node",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/sessions/{session_id}/nodes/{node_id}/reject": {
      "post": {
        "description": "The human gate's no: record the verdict, then retry with feedback.\n\n**This request is held open for the whole retry, and that is a reported\ndefect rather than a design.** C7 flagged it for C9 to decide, and the\ndecision is that the retry should be *scheduled* \u2014 the verdict is durable\nbefore the transition to ``ready``, so a process that dies between them\nrestarts into a node the scheduler picks up carrying the feedback anyway.\n\nIt is not implemented here because it cannot be implemented here honestly.\n``reject_node`` is one call that performs the state check, writes the\nverdict, transitions the node and then runs an agent; the transport has no\npoint between the fourth and the fifth at which it could answer 202. The\ntwo ways to fake it are both worse than the wait: returning 202 before the\ncheck means an invalid transition arrives as an unobserved task exception\ninstead of a 409, and pre-checking the status in this function puts the\nstate machine in a route (`docs/architecture.md` \u00a71 rule 3). The fix is a\nservice method that returns after the transition; see C9's report.",
        "operationId": "reject_node_api_sessions__session_id__nodes__node_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Node",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/sessions/{session_id}/nodes/{node_id}/retry": {
      "post": {
        "operationId": "retry_node_api_sessions__session_id__nodes__node_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RetryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Node",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/sessions/{session_id}/nodes/{node_id}/reviews": {
      "get": {
        "operationId": "list_reviews_api_sessions__session_id__nodes__node_id__reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NodeReviewResponse"
                  },
                  "title": "Response List Reviews Api Sessions  Session Id  Nodes  Node Id  Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reviews",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/sessions/{session_id}/nodes/{node_id}/runs": {
      "post": {
        "description": "Run one already-materialized node to its terminal or gated state.\n\nSynchronous, exactly as Phase 1's ``POST /runs`` is: the response *is* the\noutcome. See the module note on :func:`reject_node` for why that is a\ndeliberate cost here and a reported problem there.",
        "operationId": "run_node_api_sessions__session_id__nodes__node_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Node",
        "tags": [
          "nodes"
        ]
      }
    },
...
  }
}